api/requirements.txt
fastapi
uvicorn
python-dotenv
requests
tenacity
langdetect
together
